update: detect.pyをロボコン用に編集中
