+++ machine_learning/label_data.yaml
@@ -1,0 +1,6 @@
+names: 
+- front
+- back
+- right
+- left
+nc: 4
