--- src/detect_fig/label_data.yaml
@@ -1,6 +1,0 @@
-names: 
-- front
-- back
-- right
-- left
-nc: 4
